restructure logStore plugin files into standalone and network paths
