+++ packages/core/src/plugins/logStore/BaseQueryRequestManager.ts
@@ -5,8 +5,7 @@ import {
 	QueryRequest,
 	QueryType,
 } from '@logsn/protocol';
-import { Logger } from '@streamr/utils';
-import { Readable } from 'stream';
+import {Readable} from 'stream';
 
 import {
 	LogStore,
@@ -14,9 +13,8 @@ import {
 	MIN_SEQUENCE_NUMBER_VALUE,
 } from './LogStore';
 
-const logger = new Logger(module);
 
-export class StandaloneQueryRequestManager {
+export class BaseQueryRequestManager {
 	private logStore?: LogStore;
 
 	constructor() {
@@ -31,7 +29,7 @@ export class StandaloneQueryRequestManager {
 		let readableStream: Readable;
 		switch (queryRequest.queryType) {
 			case QueryType.Last: {
-				const { last } = queryRequest.queryOptions as QueryLastOptions;
+				const {last} = queryRequest.queryOptions as QueryLastOptions;
 
 				readableStream = this.logStore!.requestLast(
 					queryRequest.streamId,
@@ -41,7 +39,7 @@ export class StandaloneQueryRequestManager {
 				break;
 			}
 			case QueryType.From: {
-				const { from, publisherId, limit } =
+				const {from, publisherId, limit} =
 					queryRequest.queryOptions as QueryFromOptions;
 
 				readableStream = this.logStore!.requestFrom(
@@ -55,7 +53,7 @@ export class StandaloneQueryRequestManager {
 				break;
 			}
 			case QueryType.Range: {
-				const { from, publisherId, to, msgChainId, limit } =
+				const {from, publisherId, to, msgChainId, limit} =
 					queryRequest.queryOptions as QueryRangeOptions;
 
 				readableStream = this.logStore!.requestRange(
