+++ packages/core/src/plugins/logStore/LogStorePlugin.ts
@@ -43,6 +43,8 @@ export interface LogStorePluginConfig {
 	};
 }
 
+type a = Node
+
 export abstract class LogStorePlugin extends Plugin<LogStorePluginConfig> {
 	protected maybeLogStore?: LogStore;
 	protected maybeLogStoreConfig?: LogStoreConfig;
