+++ packages/core/src/plugins/logStore/http/authentication.ts
