+++ packages/core/src/plugins/logStore/standalone/BaseQueryRequestManager.ts
@@ -1,0 +1,3 @@
+import { Logger } from '@streamr/utils';
+
+const logger = new Logger(module);
